Using routing inplace of state conditional rendering
