+++ package.json
@@ -18,6 +18,7 @@
     "font-awesome": "^4.7.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.2.1",
     "react-scripts": "4.0.3",
     "reactstrap": "^8.10.0",
     "truffle": "^5.4.13",
