+++ src/Components/Styles/home.module.css
@@ -1,0 +1,67 @@
+.wrapper {
+    padding: 150px 0px 100px 0px;
+}
+.home_cover {
+    float: right;
+}
+.about_wrapper {
+    margin: 0;
+    top: 50%;
+    position: relative;
+    -ms-transform: translateY(-50%);
+    transform: translateY(-50%);
+}
+
+.title {
+    font-weight: bold;
+    font-size: 4vw;
+}
+
+.about {
+    font-weight: 500;
+    font-size: 1vw;
+    color: #6B6E70;
+}
+
+.connect {    
+    background-color: #f2f2f3;
+    border-radius: 50px;
+    padding: 3px 10px;
+}
+.connect a{
+    font-weight: bold;
+    font-size: medium;
+    text-align: center;
+    color: #61892F;
+}
+.connect a:hover{
+    color: #435f21;
+}
+@media only screen and (max-width: 768px) {
+    .home_cover {
+        float: left
+    }    
+    .about_wrapper {
+        text-align: center;
+        margin-top: 50px;
+    }
+    .title {
+        font-size: 7vw;
+    }
+    .about {
+        font-size: 3vw;
+    }
+}
+
+
+@media only screen and (max-width: 576px) {
+    .wrapper {
+        padding: 100px 0px 50px 0px;
+    }    
+    .title {
+        font-size: 8vw;
+    }
+    .about {
+        font-size: 4vw;
+    }
+}
